users-authentication 3-5

+++ app/Http/Controllers/PostController.php
@@ -18,6 +18,9 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
@@ -38,12 +41,18 @@ class PostController extends Controller
 
     public function getAdminCreate()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
@@ -72,6 +81,9 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
@@ -88,6 +100,9 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $post->likes()->delete();
         $post->tags()->detach();
